requirements.txt
absl-py==2.1.0
accelerate==1.4.0
accelerator==2024.9.13
addict==2.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
array_record==0.7.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
bddl==1.0.1
blinker==1.9.0
bottle==0.12.25
cachetools==5.5.2
certifi==2025.1.31
cffi==1.14.2
charset-normalizer==3.4.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==2.1.0
cmake==4.0.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
daqp==0.6.0
dash==3.0.4
decorator==5.2.1
deepspeed==0.16.7
dm-tree==0.1.9
docker-pycreds==0.4.0
docstring_parser==0.16
draccus==0.10.0
easydict==1.9
egl_probe==1.0.2
einops==0.8.1
etils==1.12.2
evdev==1.9.1
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flash_attn==2.7.4.post1
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
ftfy==6.3.1
future==0.18.2
fvcore==0.1.5.post20221221
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.9.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.71.0
gym==0.25.2
gym-notices==0.0.8
h5py==3.13.0
hjson==3.1.0
html-testRunner==1.2.1
huggingface-hub==0.30.2
hydra-core==1.2.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
immutabledict==4.2.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
iopath==0.1.10
ipython==8.36.0
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
jupytext==1.17.0
keras==2.15.0
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.5.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mergedeep==1.3.4
mink==0.0.10
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.1.0
mujoco==3.3.1
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.37.1
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.60.0
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.19.0
open_clip_torch==2.32.0
opencv-python==4.6.0.66
OpenEXR==3.3.3
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.5.0
portalocker==3.1.1
promise==2.3
prompt_toolkit==3.0.51
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pynput==1.8.1
PyOpenGL==3.1.9
pyparsing==3.2.3
pyquaternion==0.9.9
pytest==8.3.5
python-dateutil==2.9.0.post0
python-xlib==0.33
pytorch3d @ git+https://github.com/YanjieZe/3D-Diffusion-Policy.git@5207c41ad917684c6f2b9d2900c0d7a593df94ab#subdirectory=third_party/pytorch3d_simplified
pytz==2025.2
PyYAML==6.0.2
pyyaml-include==1.4.1
qpsolvers==4.6.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==14.0.0
robomimic==0.2.0
robosuite==1.4.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.26.0
setproctitle==1.3.5
shortuuid==1.0.13
simple-parsing==0.1.7
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.7
tensorflow-estimator==2.15.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.1
termcolor==3.0.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.6.0
torchdiffeq==0.2.5
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
trimesh==4.6.6
triton==3.2.0
typeguard==2.13.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
waitress==3.0.2
wandb==0.13.1
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wrapt==1.14.1
yacs==0.1.8
zipp==3.21.0
